+++ worker/src/types.d.ts
@@ -4,6 +4,19 @@ type UserRole = {
     prefix: string | undefined | null
 }
 
+type SmtpImapProxyConfig = {
+    smtp?: {
+        host?: string
+        port?: number | string
+        starttls?: boolean | string
+    }
+    imap?: {
+        host?: string
+        port?: number | string
+        starttls?: boolean | string
+    }
+}
+
 type Bindings = {
     // bindings
     DB: D1Database
@@ -45,6 +58,8 @@ type Bindings = {
     DISABLE_ANONYMOUS_USER_CREATE_EMAIL: string | boolean | undefined
     ENABLE_USER_DELETE_EMAIL: string | boolean | undefined
     ENABLE_ADDRESS_PASSWORD: string | boolean | undefined
+    ENABLE_AGENT_EMAIL_INFO: string | boolean | undefined
+    SMTP_IMAP_PROXY_CONFIG: string | SmtpImapProxyConfig | undefined
     ENABLE_INDEX_ABOUT: string | boolean | undefined
     DEFAULT_SEND_BALANCE: number | string | undefined
     NO_LIMIT_SEND_ROLE: string | undefined | null